# openapi.yaml
openapi: 3.0.0
paths:
...
  /api/admin/candidates:
    post:
      tags:
        - ADMIN
      summary: 'Создание студента'
      operationId: 3b17c617fad2984d0721794131f3cea3
      requestBody:
        description: 'Параметры для создания студента'
        required: true
        content:
          application/json:
            schema:
              properties:
                fio:
                  type: string
                about:
                  type: string
                email:
                  type: string
                numz:
                  type: integer
                phone:
                  type: string
                course:
                  type: integer
                training_group:
                  type: string
                canSendParticipations:
                  type: boolean
                skill_ids:
                  type: array
                  items: { type: integer }
              type: object
      responses:
        '200':
          description: 'Информация о студенте создана'
    patch:
      tags:
        - ADMIN
      summary: 'Обновить информацию о студенте @TODO BODY'
      operationId: 2863c09aad77ed86e41eb20ea249ce4f
      parameters:
        -
          name: id
          in: path
          description: 'ID студента'
          required: true
          schema:
            type: integer
      requestBody:
        description: 'Параметры для обновления студента'
        required: true
        content:
          application/json:
            schema:
              properties:
                fio:
                  type: string
                about:
                  type: string
                email:
                  type: string
                numz:
                  type: integer
                phone:
                  type: string
                course:
                  type: integer
                training_group:
                  type: string
                skill_ids:
                  type: array
                  items: { type: integer }
              type: object
      responses:
        '200':
          description: 'Информация о студенте обновлена'
        '404':
          description: 'Студент не найден'
...
  /api/admin/participations:
    get:
      tags:
        - ADMIN
      summary: 'Получить все заявки'
      operationId: 08cfd783e97365a1e9560a470fdc89f9
      responses:
        '200':
          description: 'Список всех заявок'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Participation'
    post:
      tags:
        - ADMIN
      summary: 'Создать заявку @TODO BODY'
      operationId: 16176789c58375bad7d2c751c15b2b38
      requestBody:
        description: 'Параметры для создания заявки'
        required: true
        content:
          application/json:
            schema:
              properties:
                priority:
                  type: integer
                state_id:
                  type: integer
                candidate_id:
                  type: integer
                project_id:
                  type: integer
              type: object
      responses:
        '200':
          description: 'Заявка создана'
        '400':
          description: 'У кандидата уже есть заявка на этот проект'
...
  '/api/admin/projects/${id}':
    delete:
      tags:
        - ADMIN
      summary: 'Удаление проекта'
      operationId: 3e5ccd2e6e3d9bd429b36cae215713cf
      parameters:
        -
          name: id
          in: path
          description: 'ID проекта'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Проект удален'
        '404':
          description: 'Проект не найден'
    patch:
      tags:
        - ADMIN
      summary: 'Обновить проект'
      operationId: e2c4a11d12a4215787f8f46b11c805d8
      parameters:
        -
          name: id
          in: path
          description: 'ID проекта'
          required: true
          schema:
            type: integer
      requestBody:
        description: 'Параметры для обновления проекта'
        required: true
        content:
          application/json:
            schema:
              properties:
                title:
                  type: string
                places:
                  type: integer
                goal:
                  type: string
                description:
                  type: string
                difficulty:
                  type: integer
                date_start:
                  type: string
                date_end:
                  type: string
                requirements:
                  type: string
                customer:
                  type: string
                product_result:
                  type: string
                study_result:
                  type: string
                additional_inf:
                  type: string
                state_id:
                  type: integer
                type_id:
                  type: integer
                supervisor_ids:
                  type: array
                  items: { type: integer }
                skill_ids:
                  type: array
                  items: { type: integer }
                speciality_ids:
                  type: array
                  items: { type: integer }
              type: object
      responses:
        '200':
          description: 'Проект изменен'
        '404':
          description: 'Проект не найден'
...
  /api/admin/projects:
    post:
      tags:
        - ADMIN
      summary: 'Создать проект'
      operationId: 111fb580612eb7b269c09a8bc49db764
      requestBody:
        description: 'Параметры для создания проекта'
        required: true
        content:
          application/json:
            schema:
              properties:
                title:
                  type: string
                places:
                  type: integer
                goal:
                  type: string
                description:
                  type: string
                difficulty:
                  type: integer
                date_start:
                  type: string
                  format: date
                date_end:
                  type: string
                  format: date
                requirements:
                  type: string
                customer:
                  type: string
                product_result:
                  type: string
                study_result:
                  type: string
                additional_inf:
                  type: string
                state_id:
                  type: integer
                type_id:
                  type: integer
                supervisor_ids:
                  type: array
                  items: { type: integer }
                skill_ids:
                  type: array
                  items: { type: integer }
                speciality_ids:
                  type: array
                  items: { type: integer }
              type: object
      responses:
        '200':
          description: 'Проект создан'
...
  /api/admin/states:
    post:
      tags:
        - ADMIN
      summary: 'Создать состояние проекта'
      operationId: 0c942b0a43a4437c869b5940e20f26e3
      requestBody:
        description: 'Параметры для обновления состояния проекта'
        required: true
        content:
          application/json:
            schema:
              properties:
                state:
                  type: string
              type: object
      responses:
        '200':
          description: 'Состояние проекта создано'
...
components:
  schemas:
...
    Participation:
      properties:
        id:
          description: 'ID заявки'
          type: integer
        priority:
          description: 'Приоритет заявки'
          type: integer
        project_id:
          description: 'ID проекта, на которые подали заявку'
          type: string
        state_id:
          description: 'ID состояния'
          type: string
        candidate_id:
          description: 'ID студента подавшего заявку'
          type: string
        created_at:
          description: 'Дата создания записи в БД'
          type: string
        updated_at:
          description: 'Дата обновления записи в БД'
          type: string
      type: object
...